Update models and move them to correct namespace
