+++ js/src/domains/compareCustomize/components/maps/InterventionPlanMap.tsx
@@ -2,10 +2,10 @@ import React, { FC, useCallback, useMemo } from 'react';
 import { Box } from '@mui/material';
 import { SxStyles } from 'Iaso/types/general';
 import { Map as SNTMap } from '../../../../components/Map';
+import { Intervention } from '../../../interventions/types';
 import { useGetAccountSettings } from '../../../planning/hooks/useGetAccountSettings';
 import { useGetInterventionPlans } from '../../../planning/hooks/useGetInterventionAssignments';
 import { useGetOrgUnits } from '../../../planning/hooks/useGetOrgUnits';
-import { Intervention } from '../../../planning/types/interventions';
 import { NO_INTERVENTION_COLOR } from '../../utils/colors';
 import {
     buildOrgUnitLookup,
@@ -64,8 +64,7 @@ export const InterventionPlanMap: FC<Props> = ({
     titleText,
 }) => {
     const { data: accountSettings } = useGetAccountSettings();
-    const interventionTypeId =
-        accountSettings?.intervention_org_unit_type_id;
+    const interventionTypeId = accountSettings?.intervention_org_unit_type_id;
     const { data: orgUnits } = useGetOrgUnits({
         orgUnitTypeId: interventionTypeId,
         enabled: !!interventionTypeId,
