+++ js/src/domains/interventions/types/interventionCategories.ts
@@ -1,0 +1,9 @@
+import { Intervention } from './intervention';
+
+export type InterventionCategory = {
+    id: number;
+    name: string;
+    short_name: string;
+    description: string;
+    interventions: Intervention[];
+};
